requirements.txt
pillow==9.5
opencv-python
pytube
